WIP: media: dvb-frontends: add WeTek Play2 multi-standard device
This is a hacky attempt to understand more of the plumbing needed to
get embedded DVB cards working. I have borderline no clue what I'm
doing here so if you found this commit and got excited about support
for the dvb tuners in your box, please adjust expectations to zero!

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>