HACK: arm64: dts: meson: add p261/p271 support
Add device-trees for Amlogic P261/P271 (S905L) reference design boards. Similar to the
P281 (S905W) the boards are derived from P212 (S905X) but with silicon differences:

- P271 omits VP9 support and uses Mali 450-MP3
- P261 omits VP9 support and uses Mali 450-MP2

P271 boards have an S905L SoC with "B-3" or "3" marking. P261 boards have an S905L SoC
with "B-2" or "2" marking. The B is believed to denote a chip with DTS license and the
number is believed to reference the Mali 450 core(s) spec. The MP2 variant is referred
to as "S905C" in some Amlogic marketing presos and vendor code, and "S905L2" on chip
distributor stock lists.

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>